Move getConnectedJetpackInfo to .~/utils/
Address https://github.com/woocommerce/google-listings-and-ads/pull/893#discussion_r674034773
