+++ js/src/data/utils.js
@@ -3,7 +3,6 @@
  */
 import { format } from '@wordpress/date';
 import { getCurrentDates } from '@woocommerce/date';
-import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
@@ -166,34 +165,6 @@ export function calculateDelta( value, base ) {
 	}
 
 	return delta;
-}
-
-/**
- * @typedef {import('./selectors').JetpackAccount} JetpackAccount
- */
-/**
- * Get a connected jetpack information for display on UI.
- * This function should always be called after the jetpack is connected.
- *
- * @param {JetpackAccount} jetpack Connected jetpack account data.
- * @return {string} Return email if current user is jetpack owner.
- *                  Otherwise, return a successfully connected info only.
- */
-export function getConnectedJetpackInfo( jetpack ) {
-	// Error-proofing for calling with not yet connect jetpack.
-	if ( jetpack.active !== 'yes' ) {
-		return '';
-	}
-
-	if ( jetpack.owner === 'yes' ) {
-		return jetpack.email;
-	}
-
-	// Show a connected text instead of owner email to non jetpack owner users.
-	return __(
-		'Successfully connected through Jetpack',
-		'google-listings-and-ads'
-	);
 }
 
 /**
