+++ js/src/data/utils.test.js
@@ -6,7 +6,6 @@ import {
 	getReportKey,
 	calculateDelta,
 	mapReportFieldsToPerformance,
-	getConnectedJetpackInfo,
 	freeFields,
 	paidFields,
 	MISSING_FREE_LISTINGS_DATA,
@@ -261,37 +260,6 @@ describe( 'mapReportFieldsToPerformance', () => {
 			a: {
 				missingFreeListingsData: MISSING_FREE_LISTINGS_DATA.FOR_REQUEST,
 			},
-		} );
-	} );
-} );
-
-describe( 'getConnectedJetpackInfo', () => {
-	it( 'When jetpack in not connected, should return an empty string', () => {
-		expect( getConnectedJetpackInfo( { active: 'no' } ) ).toBe( '' );
-	} );
-
-	describe( 'For connected jetpack', () => {
-		it( 'When current user is the jetpack owner, should return email', () => {
-			const jetpack = {
-				active: 'yes',
-				owner: 'yes',
-				email: 'i_am_jetpack_owner@example.com',
-			};
-
-			expect( getConnectedJetpackInfo( jetpack ) ).toBe( jetpack.email );
-		} );
-
-		it( 'When current user is not the jetpack owner, should return a successfully connected info', () => {
-			const jetpack = {
-				active: 'yes',
-				owner: 'no',
-				email: 'i_am_another_admin_user@example.com',
-			};
-
-			const info = getConnectedJetpackInfo( jetpack );
-
-			expect( info ).not.toBe( jetpack.email );
-			expect( info ).toMatchSnapshot();
 		} );
 	} );
 } );
